backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.12
httpx==0.27.2
python-dotenv==1.0.1

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0

# Phase 2 ingestion dependencies.
#
# These are imported lazily by the ingestion stages. The pipeline still
# runs end-to-end without them (text uploads work; PDF uploads degrade
# gracefully and report the missing backend), so they are listed here for
# production installs and CI rather than as hard requirements.
PyMuPDF==1.24.10
pdfplumber==0.11.4
pytesseract==0.3.13
Pillow==10.4.0

# Optional providers — uncomment to enable.
# openai==1.51.0
# anthropic==0.34.2
